Tuya MCU integrations for Yesncer YG-20 and YG-30 FCUs
These are using the Tuya WBR3 WiFi modules that are already in the
thermostat displays, which is nice because it requires no hardware
changes, but unfortunately the LibreTiny integration for them is not
even released yet and in my experience from running this for a few
weeks is that it is not stable yet. If the WiFi/API connection is
lost, it seems to not be recovered and you have to power-cycle the
FCU.
Also, the current lack of OTA support is very unfortunate since you
have to desolder the module from the PCB in order to flash it.

I plan to make a second version of this soon which will replace the
WBR3 module with an ESP and make use of the Modbus integration in
addition to Tuya MCU.

This is currently still using the legacy config inheritance format
instead of the much nicer packages. Given that I will be replacing
these soon anyway, I'm not going to switch them over.