Vitocal 060-A integration
This makes use of a UART debug interface on the Cotherm IFH90005
thermostat the unit uses, which spits out a string message with a
bunch of hex fields every second. My father and I have managed to
reverse-engineer most of them. This integration has been running
without issues for more than half a year now.

I'll clean the config up (migrate to packages mostly) and properly
document it in the future.